new chars, notify setup

+++ VMS_Device/app/accelerometer/inc/acc.h
@@ -5,9 +5,9 @@
 
 typedef struct
 {
-	float x_axis_val;
-	float y_axis_val;
-	float z_axis_val;
+	uint32_t x_axis_val;
+	uint32_t y_axis_val;
+	uint32_t z_axis_val;
 }acc_meas;
 
 extern acc_meas acc_meas_values;
